requirements.txt
# Core
aiogram>=3.4.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0
pytz>=2024.1

# Web panel
aiohttp>=3.9.0

# Optional: Redis for FSM storage (recommended for production)
redis>=5.0.0
